+++ primitives/src/block.rs
@@ -7,9 +7,83 @@
 //! module describes structures and functions needed to describe
 //! these blocks and the blockchain.
 
+use core::fmt;
+
 #[cfg(feature = "arbitrary")]
 use arbitrary::{Arbitrary, Unstructured};
-use hashes::sha256d;
+use hashes::{sha256d, HashEngine as _};
+
+use crate::merkle_tree::TxMerkleNode;
+use crate::pow::CompactTarget;
+
+/// Bitcoin block header.
+///
+/// Contains all the block's information except the actual transactions, but
+/// including a root of a [Merkle tree] committing to all transactions in the block.
+///
+/// [Merkle tree]: https://en.wikipedia.org/wiki/Merkle_tree
+///
+/// ### Bitcoin Core References
+///
+/// * [CBlockHeader definition](https://github.com/bitcoin/bitcoin/blob/345457b542b6a980ccfbc868af0970a6f91d1b82/src/primitives/block.h#L20)
+#[derive(Copy, PartialEq, Eq, Clone, PartialOrd, Ord, Hash)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+pub struct Header {
+    /// Block version, now repurposed for soft fork signalling.
+    pub version: Version,
+    /// Reference to the previous block in the chain.
+    pub prev_blockhash: BlockHash,
+    /// The root hash of the Merkle tree of transactions in the block.
+    pub merkle_root: TxMerkleNode,
+    /// The timestamp of the block, as claimed by the miner.
+    pub time: u32,
+    /// The target value below which the blockhash must lie.
+    pub bits: CompactTarget,
+    /// The nonce, selected to obtain a low enough blockhash.
+    pub nonce: u32,
+}
+
+impl Header {
+    /// The number of bytes that the block header contributes to the size of a block.
+    // Serialized length of fields (version, prev_blockhash, merkle_root, time, bits, nonce)
+    pub const SIZE: usize = 4 + 32 + 32 + 4 + 4 + 4; // 80
+
+    /// Returns the block hash.
+    // This is the same as `Encodable` but done manually because `Encodable` isn't in `primitives`.
+    pub fn block_hash(&self) -> BlockHash {
+        let mut engine = sha256d::Hash::engine();
+        engine.input(&self.version.to_consensus().to_le_bytes());
+        engine.input(self.prev_blockhash.as_byte_array());
+        engine.input(self.merkle_root.as_byte_array());
+        engine.input(&self.time.to_le_bytes());
+        engine.input(&self.bits.to_consensus().to_le_bytes());
+        engine.input(&self.nonce.to_le_bytes());
+
+        BlockHash::from_byte_array(sha256d::Hash::from_engine(engine).to_byte_array())
+    }
+}
+
+impl fmt::Debug for Header {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.debug_struct("Header")
+            .field("block_hash", &self.block_hash())
+            .field("version", &self.version)
+            .field("prev_blockhash", &self.prev_blockhash)
+            .field("merkle_root", &self.merkle_root)
+            .field("time", &self.time)
+            .field("bits", &self.bits)
+            .field("nonce", &self.nonce)
+            .finish()
+    }
+}
+
+impl From<Header> for BlockHash {
+    fn from(header: Header) -> BlockHash { header.block_hash() }
+}
+
+impl From<&Header> for BlockHash {
+    fn from(header: &Header) -> BlockHash { header.block_hash() }
+}
 
 /// Bitcoin block version number.
 ///
@@ -91,6 +165,20 @@ hashes::hash_newtype! {
 impl BlockHash {
     /// Dummy hash used as the previous blockhash of the genesis block.
     pub const GENESIS_PREVIOUS_BLOCK_HASH: Self = Self::from_byte_array([0; 32]);
+}
+
+#[cfg(feature = "arbitrary")]
+impl<'a> Arbitrary<'a> for Header {
+    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
+        Ok(Header {
+            version: Version::arbitrary(u)?,
+            prev_blockhash: BlockHash::from_byte_array(u.arbitrary()?),
+            merkle_root: TxMerkleNode::from_byte_array(u.arbitrary()?),
+            time: u.arbitrary()?,
+            bits: CompactTarget::from_consensus(u.arbitrary()?),
+            nonce: u.arbitrary()?,
+        })
+    }
 }
 
 #[cfg(feature = "arbitrary")]
